+++ final-project-focus00/focus/src/main/java/com/example/focus/Service/ShiftService.java
@@ -1,8 +1,10 @@
 package com.example.focus.Service;
 
+import com.example.focus.ApiResponse.ApiException;
 import com.example.focus.DTO.ShiftDTO;
 import com.example.focus.DTO.ShiftDTOIn;
 import com.example.focus.Model.Shift;
+import com.example.focus.Model.Space;
 import com.example.focus.Repository.ShiftRepository;
 import com.example.focus.Repository.SpaceRepository;
 import lombok.RequiredArgsConstructor;
@@ -29,29 +31,43 @@ public class ShiftService {
         return shift;
     }
 
-    public Shift createShift(ShiftDTOIn shiftDTOIn) {
+    public Shift createShift(ShiftDTOIn shiftDTOIn,Integer authId) {
         Shift shift = new Shift();
+        Space space=spaceRepository.findSpaceById(shiftDTOIn.getSpaceId());
+        if (!space.getStudio().getMyUser().getId().equals(authId)){
+            throw new ApiException("you are not authorized to add this shift to this space");
+        }
         shift.setName(shiftDTOIn.getName());
-        shift.setSpace(spaceRepository.findSpaceById(shiftDTOIn.getSpaceId()));
+        shift.setSpace(space);
         shift.setStartTime(shiftDTOIn.getStartTime());
         shift.setEndTime(shiftDTOIn.getEndTime());
         shift.setStatus("Available");
         return shiftRepository.save(shift);
     }
 
-    public Shift updateShiftStatus(Integer id, String status) {
+    public Shift updateShiftStatus(Integer id, String status,Integer authId) {
         Shift shift = shiftRepository.findById(id).orElse(null);
         if (shift == null) {
             throw new IllegalArgumentException("Shift not found");
         }
+        Space space=spaceRepository.findSpaceById(shift.getSpace().getId());
+        if (!space.getStudio().getMyUser().getId().equals(authId)){
+            throw new ApiException("you are not authorized to update this shift for this space");
+        }
+
         shift.setStatus(status);
         return shiftRepository.save(shift);
     }
 
-    public void deleteShift(Integer id) {
+    public void deleteShift(Integer id,Integer authId) {
+
         Shift shift = shiftRepository.findById(id).orElse(null);
         if (shift == null) {
-            throw new IllegalArgumentException("Shift not found");
+            throw new ApiException("Shift not found");
+        }
+        Space space=spaceRepository.findSpaceById(shift.getSpace().getId());
+        if (!space.getStudio().getMyUser().getId().equals(authId)){
+            throw new ApiException("you are not authorized to delete this shift from this space");
         }
         shiftRepository.delete(shift);
     }
